+++ src/Dojang/Commands/Init.hs
@@ -13,6 +13,7 @@ module Dojang.Commands.Init
   ) where
 
 import Control.Monad (forM_, when)
+import Control.Monad.Except (MonadError (throwError))
 import Control.Monad.IO.Class (MonadIO (liftIO))
 import Data.Function ((&))
 import Data.List (maximumBy, sortOn)
@@ -357,7 +358,9 @@ init presets noInteractive = do
   debug' <- asks (.debug)
   dryRun' <- asks (.dryRun)
   when (debug' || dryRun') $ do
-    let manifestText = indent $ writeManifest manifest
+    manifestText <- case writeManifest manifest of
+      Left err -> throwError $ userError $ show err
+      Right toml -> pure $ indent toml
     printStderr' Note $
       "The manifest file looks like below:\n\n"
         <> manifestText