+++ src/Dojang/Syntax/Manifest/Writer.hs
@@ -3,14 +3,16 @@
 {-# LANGUAGE NoFieldSelectors #-}
 
 module Dojang.Syntax.Manifest.Writer
-  ( writeManifest
+  ( WriteError (..)
+  , writeManifest
   , writeManifestFile
   ) where
 
-import Data.Bifunctor (Bifunctor (bimap, second))
-import Data.List (partition)
+import Control.Monad.Except (MonadError (throwError))
+import Data.Bifunctor (Bifunctor (second))
+import Data.List (group, partition, sort)
 import Data.List.NonEmpty (NonEmpty ((:|)), length, toList)
-import Data.Maybe (catMaybes, fromMaybe, listToMaybe)
+import Data.Maybe (fromMaybe, listToMaybe)
 import GHC.IsList (IsList (fromList))
 import System.IO.Unsafe (unsafePerformIO)
 import Prelude hiding (all, any, writeFile)
@@ -24,7 +26,6 @@ import System.OsPath (OsPath, decodeFS)
 import TextShow (FromStringShow (FromStringShow), TextShow (showt))
 import Toml.Pretty (prettyTomlOrdered)
 import Toml.ToValue (ToTable (toTable))
-import Toml.Value (Table)
 
 import qualified Data.Map.Strict
 import Dojang.MonadFileSystem
@@ -48,7 +49,7 @@ import Dojang.Types.EnvironmentPredicate
   ( EnvironmentPredicate (..)
   , normalizePredicate
   )
-import Dojang.Types.FilePathExpression (toPathText)
+import Dojang.Types.FilePathExpression (FilePathExpression, toPathText)
 import Dojang.Types.FileRoute (FileRoute (..))
 import Dojang.Types.FileRouteMap (FileRouteMap)
 import Dojang.Types.Hook (Hook (..), HookMap, HookType (..))
@@ -61,18 +62,56 @@ schema :: Text
 schema = "https://schema.dojang.dev/2023-11/manifest.schema.json"
 
 
+-- | An error that can occur while encoding a 'Manifest'.
+data WriteError
+  = -- | A route contains an environment predicate that cannot be represented
+    -- by any moniker in the manifest.
+    UnrepresentableRoutePredicate
+      OsPath
+      -- ^ The source path of the route.
+      EnvironmentPredicate
+      -- ^ The predicate that cannot be represented.
+  | -- | Multiple route predicates resolve to the same moniker and therefore
+    -- cannot be represented as distinct entries in the manifest.
+    DuplicateRouteMoniker
+      OsPath
+      -- ^ The source path of the route.
+      MonikerName
+      -- ^ The moniker shared by multiple route predicates.
+  deriving (Eq)
+
+
+instance Show WriteError where
+  show (UnrepresentableRoutePredicate path predicate) =
+    "Route predicate "
+      <> show predicate
+      <> " for path "
+      <> show (decodePath path)
+      <> " cannot be represented by any moniker."
+  show (DuplicateRouteMoniker path moniker) =
+    "Multiple route predicates for path "
+      <> show (decodePath path)
+      <> " resolve to moniker "
+      <> show moniker
+      <> "."
+
+
 -- | Encodes a 'Manifest' into a TOML document.
-writeManifest :: Manifest -> Text
-writeManifest manifest =
-  "#:schema "
-    <> schema
-    <> "\n\n"
-    <> (showt $ FromStringShow $ prettyTomlOrdered order tbl)
+writeManifest
+  :: Manifest
+  -- ^ The 'Manifest' to encode.
+  -> Either WriteError Text
+  -- ^ The encoded TOML document, or an error if the manifest contains route
+  -- predicates that the TOML format cannot represent without data loss.
+writeManifest manifest = do
+  manifest' <- mapManifest' manifest
+  let tbl = toTable manifest'
+  pure $
+    "#:schema "
+      <> schema
+      <> "\n\n"
+      <> (showt $ FromStringShow $ prettyTomlOrdered order tbl)
  where
-  manifest' :: Manifest'
-  manifest' = mapManifest' manifest
-  tbl :: Table
-  tbl = toTable manifest'
   order :: [String] -> String -> Either Int String
   order [] field = case field of
     "dirs" -> Left 1
@@ -83,8 +122,8 @@ writeManifest manifest =
   order _ field = Right field
 
 
--- | Writes a 'Manifest' file to the given path.  Throw an 'IOError' if the
--- an occurs while writing the file.
+-- | Writes a 'Manifest' file to the given path.  Throws an 'IOError' if the
+-- manifest cannot be encoded or an error occurs while writing the file.
 writeManifestFile
   :: (MonadFileSystem m)
   => Manifest
@@ -93,16 +132,16 @@ writeManifestFile
   -- ^ The path to write the 'Manifest' to.
   -> m ()
 writeManifestFile manifest filePath =
-  writeFile filePath $ encodeUtf8 $ writeManifest manifest
+  case writeManifest manifest of
+    Left err -> throwError $ userError $ show err
+    Right toml -> writeFile filePath $ encodeUtf8 toml
 
 
-mapManifest' :: Manifest -> Manifest'
-mapManifest' manifest =
-  Manifest' monikers' dirs files ignores hooks'
+mapManifest' :: Manifest -> Either WriteError Manifest'
+mapManifest' manifest = do
+  (dirs, files) <- mapFiles manifest.fileRoutes manifest.monikers
+  pure $ Manifest' monikers' dirs files ignores hooks'
  where
-  dirs :: FileRouteMap'
-  files :: FileRouteMap'
-  (dirs, files) = mapFiles manifest.fileRoutes manifest.monikers
   monikers' :: MonikerMap'
   monikers' = mapMonikers' manifest.monikers
   ignores :: IgnoreMap'
@@ -118,12 +157,19 @@ mapManifest' manifest =
       else Just $ mapHooks' manifest.hooks
 
 
-mapFiles :: FileRouteMap -> MonikerMap -> (FileRouteMap', FileRouteMap')
-mapFiles fileRouteMap monikers =
-  ( fromList (bimap decodePath (`mapFileRoute'` monikers) <$> dirs)
-  , fromList (bimap decodePath (`mapFileRoute'` monikers) <$> files)
-  )
+mapFiles
+  :: FileRouteMap
+  -> MonikerMap
+  -> Either WriteError (FileRouteMap', FileRouteMap')
+mapFiles fileRouteMap monikers = do
+  dirs' <- traverse mapRoute dirs
+  files' <- traverse mapRoute files
+  pure (fromList dirs', fromList files')
  where
+  mapRoute :: (OsPath, FileRoute) -> Either WriteError (FilePath, FileRoute')
+  mapRoute (path, route) = do
+    route' <- mapFileRoute' path route monikers
+    pure (decodePath path, route')
   dirs :: [(OsPath, FileRoute)]
   files :: [(OsPath, FileRoute)]
   (dirs, files) =
@@ -136,25 +182,42 @@ decodePath :: OsPath -> FilePath
 decodePath = unsafePerformIO . decodeFS
 
 
-mapFileRoute' :: FileRoute -> MonikerMap -> FileRoute'
-mapFileRoute' fileRoute monikers =
-  (fromList . catMaybes)
-    [ case (normalizePredicate pred', maybe "" toPathText filePath) of
-        (Moniker n, path) -> Just (n, path)
-        (pred'', path) -> case lookBack monikers pred'' of
-          Just n -> Just (n, path)
-          Nothing -> Nothing
-    | (pred', filePath) <- fileRoute.predicates
-    ]
+mapFileRoute'
+  :: OsPath -> FileRoute -> MonikerMap -> Either WriteError FileRoute'
+mapFileRoute' routePath fileRoute monikers = do
+  mappedPredicates <- traverse mapPredicate fileRoute.predicates
+  case duplicateMoniker mappedPredicates of
+    Nothing -> Right $ fromList mappedPredicates
+    Just moniker -> Left $ DuplicateRouteMoniker routePath moniker
+ where
+  duplicateMoniker :: [(MonikerName, Text)] -> Maybe MonikerName
+  duplicateMoniker entries =
+    listToMaybe
+      [ moniker
+      | moniker : _ : _ <- group $ sort $ fst <$> entries
+      ]
+  mapPredicate
+    :: (EnvironmentPredicate, Maybe FilePathExpression)
+    -> Either WriteError (MonikerName, Text)
+  mapPredicate (predicate, filePath) =
+    case normalizePredicate predicate of
+      Moniker name -> Right (name, path)
+      normalizedPredicate -> case lookBack monikers normalizedPredicate of
+        Just name -> Right (name, path)
+        Nothing ->
+          Left $ UnrepresentableRoutePredicate routePath normalizedPredicate
+   where
+    path = maybe "" toPathText filePath
 
 
 lookBack :: MonikerMap -> EnvironmentPredicate -> Maybe MonikerName
 lookBack monikers predicate =
-  listToMaybe
-    [ n
-    | (n, p) <- Data.HashMap.Strict.toList monikers
-    , normalizePredicate p == normalizedPred
-    ]
+  listToMaybe $
+    sort
+      [ n
+      | (n, p) <- Data.HashMap.Strict.toList monikers
+      , normalizePredicate p == normalizedPred
+      ]
  where
   normalizedPred :: EnvironmentPredicate
   normalizedPred = normalizePredicate predicate